Began header + Reformat Quote + Continue C2A
1. Changed header formatting to a container-1, that seperated the text
and button from the image. Setting this to "display:flex;" made it much
closer to completion.

2. Added more padding to the width of the Quote block to more closely
align with the sample image. Changed the author font from bold to normal
and tried resizing it on the page.

3. Changed the C2A to have an additional container div called
"action-text". This allowed the button, when set to flex-basis
to finally sit on the right side of the text instead of the bottom. I
also changed the font sizes and boldness' to more closely align with
sample image. I changed the sign-up button to be a solid line border,
with matching white color. But it needs padding from text to border.